extract parser functions for easier testing possibilities
